+++ Sources/CodexBar/UsageStore+Accessors.swift
@@ -60,6 +60,8 @@ extension UsageStore {
             return ZaiSettingsError.missingToken.errorDescription
         case .openrouter:
             return OpenRouterSettingsError.missingToken.errorDescription
+        case .anyrouter:
+            return AnyRouterUsageError.missingCredentials.errorDescription
         case .crossmodel:
             return CrossModelSettingsError.missingToken.errorDescription
         case .clawrouter: