+++ Sources/CodexBarCore/Config/ProviderConfigEnvironment.swift
@@ -147,6 +147,8 @@ public enum ProviderConfigEnvironment {
             SyntheticSettingsReader.apiKeyKey
         case .openrouter:
             OpenRouterSettingsReader.envKey
+        case .anyrouter:
+            AnyRouterSettingsReader.apiKeyEnvironmentKey
         case .elevenlabs:
             ElevenLabsSettingsReader.apiKeyEnvironmentKey
         case .moonshot: