+++ Sources/CodexBarCore/Providers/Providers.swift
@@ -62,6 +62,7 @@ public enum UsageProvider: String, CaseIterable, Sendable, Codable {
     case clawrouter
     case sub2api
     case wayfinder
+    case anyrouter
     case zenmux
 }
 
@@ -125,6 +126,7 @@ public enum IconStyle: String, Sendable, CaseIterable {
     case clawrouter
     case sub2api
     case wayfinder
+    case anyrouter
     case zenmux
     case combined
 }